Menu formatting
remove variable RadialQuickMenu, it was intended to control if the radial quick menu feature was enabled. Deprechiated, user can disable by not assigning a button to it.

Updated the UEVR script menu entries for fewer lines. The functions that great graphical elements were updated to have a no new line parameter.

Darker darks disables automatically if manage lighting is not checked.