In game menu B and X fix.
Resolve the B and X being swapping while the user is in any menu. Out of menu this issue has been resolved on a previous commit.

Previous commit update UEHelper.lua to swap the B and X dections in the UpdateInput function. This was done unitentionally via AI. It would cause too many issues in other scripts. Returned to the incorrect previous state.